requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
python-dotenv==1.0.0
pymongo==4.3.3
jsonschema==4.17.3
PyJWT==2.7.0
requests==2.31.0
bcrypt==4.0.1